Add profile README.MD with introduction and contact info
I have added a new profile README.MD file to my repository, which provides an introduction to myself, "@mirasu-se", and my interests, as well as my contact information. In the README.MD file, I have listed the technologies I'm interested in, what I'm currently learning, and what I'm looking to collaborate on. Additionally, I have included a LinkedIn button that links to my profile page. With this commit, I aim to give more context about myself and make it easier for others to connect and collaborate with me.